fix: circular dep preventing options sharing
