+++ packages/cli-platform-android/src/commands/buildAndroid/startServerInNewWindow.ts
@@ -1,0 +1,81 @@
+import path from 'path';
+import fs from 'fs';
+import execa from 'execa';
+import {logger} from '@react-native-community/cli-tools';
+
+export function startServerInNewWindow(
+  port: number,
+  terminal: string,
+  reactNativePath: string,
+) {
+  /**
+   * Set up OS-specific filenames and commands
+   */
+  const isWindows = /^win/.test(process.platform);
+  const scriptFile = isWindows
+    ? 'launchPackager.bat'
+    : 'launchPackager.command';
+  const packagerEnvFilename = isWindows ? '.packager.bat' : '.packager.env';
+  const portExportContent = isWindows
+    ? `set RCT_METRO_PORT=${port}`
+    : `export RCT_METRO_PORT=${port}`;
+
+  /**
+   * Set up the `.packager.(env|bat)` file to ensure the packager starts on the right port.
+   */
+  const launchPackagerScript = path.join(
+    reactNativePath,
+    `scripts/${scriptFile}`,
+  );
+
+  /**
+   * Set up the `launchpackager.(command|bat)` file.
+   * It lives next to `.packager.(bat|env)`
+   */
+  const scriptsDir = path.dirname(launchPackagerScript);
+  const packagerEnvFile = path.join(scriptsDir, packagerEnvFilename);
+  const procConfig: execa.SyncOptions = {cwd: scriptsDir};
+
+  /**
+   * Ensure we overwrite file by passing the `w` flag
+   */
+  fs.writeFileSync(packagerEnvFile, portExportContent, {
+    encoding: 'utf8',
+    flag: 'w',
+  });
+
+  if (process.platform === 'darwin') {
+    try {
+      return execa.sync(
+        'open',
+        ['-a', terminal, launchPackagerScript],
+        procConfig,
+      );
+    } catch (error) {
+      return execa.sync('open', [launchPackagerScript], procConfig);
+    }
+  }
+  if (process.platform === 'linux') {
+    try {
+      return execa.sync(terminal, ['-e', `sh ${launchPackagerScript}`], {
+        ...procConfig,
+        detached: true,
+      });
+    } catch (error) {
+      // By default, the child shell process will be attached to the parent
+      return execa.sync('sh', [launchPackagerScript], procConfig);
+    }
+  }
+  if (/^win/.test(process.platform)) {
+    // Awaiting this causes the CLI to hang indefinitely, so this must execute without await.
+    return execa('cmd.exe', ['/C', launchPackagerScript], {
+      ...procConfig,
+      detached: true,
+      stdio: 'ignore',
+    });
+  }
+  logger.error(
+    `Cannot start the packager. Unknown platform ${process.platform}`,
+  );
+  return;
+}
